Added scrypt-jane for scrypt support
This commit adds scrypt-jane as a submodule, and incorporates it source
into the project.

In order to have properly encrypted data files, we need to incorporate
some form of key-derivation. Writing one's own is usually dangerous, and
sodium doesn't provide that sort of functionality, so adding a second
crypto dependency makes sense here.