Moved and simplified the Profile class
This commit moves the Profile class out of Core, and simplifies its
functions.

Having the Profile class encapsulated within Core was visibly becoming a
poor decision - it was going to create problems in the longer run with
the login dialog. As such, this commit moves it to its own top-level
file. Profile has also been simplified to work just with the Tox
Messenger save data, and no longer with actual Tox instances & the
tox_load/save() functions.